requirements.txt
feedparser
pybars3
python-dotenv
deepgram-sdk==2.12.0
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
markdown
google-genai